Move Unit fields to AbstractUnit
